클라 인벤 버그 수정 완료
합치기와 나누기에 weightUI의 반응성을 수정
weigthUI의 매커니즘 자체를 수정
각종 버그 수정
나누기 모드에서 합치기가 가능할경우에 나누기 인터페이스에서 수량을 정하면 수량만큼 merge패킷을 보냄

+++ GSO_Client/Assets/Scripts/Managers/Contents/ObjectManager.cs
@@ -131,7 +131,7 @@ public class ObjectManager
             Box boxScr = go.GetComponent<Box>();
             boxScr.objectId = info.ObjectId;
             boxScr.interactType = InteractType.InventoryObj;
-            boxScr.SetBox(info.Box.X, info.Box.Y, info.Box.Weight);
+            boxScr.SetBox((int)info.Box.X, (int)info.Box.Y, info.Box.Weight);
             //boxScr.interactRange = 나중에 필요시 추가(박스의 종류를 나눌경우)
 
             //Add로 인벤 데이터를 생성하여 boxScr.invenData에 넣기
